fixing deployment cors error
